Update README.md
In early 2023 we tried to run a test on Colab to verify test results. Colab no longer supports TensorFlow 1.x  code. We were able to set up a virtual directory using Anaconda and run our test. Running this code on Colab now would require porting it to a newer version (2.x) of TensorFlow.